Move tck tests to projection folder
