+++ packages/graphql/tests/api-v6/tck/projection/types/temporals.test.ts
@@ -17,8 +17,8 @@
  * limitations under the License.
  */
 
-import { Neo4jGraphQL } from "../../../../src";
-import { formatCypher, formatParams, translateQuery } from "../../../tck/utils/tck-test-utils";
+import { Neo4jGraphQL } from "../../../../../src";
+import { formatCypher, formatParams, translateQuery } from "../../../../tck/utils/tck-test-utils";
 
 describe("Temporal types", () => {
     let typeDefs: string;
